+++ src/renderer/hooks/index.ts
@@ -1,2 +1,4 @@
 export { useResourceEvents } from './useResourceEvents'
 export { useYAMLEditor } from './useYAMLEditor'
+export { useLogBuffer } from './useLogBuffer'
+export { useAutoScroll } from './useAutoScroll'